refactor: update imports to use shadcn directory in materials pages
