+++ app/users/dashboard/resources/materials/[id]/edit/page.tsx
@@ -9,11 +9,11 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Textarea } from '@/components/ui/textarea';
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
+import { Input } from '@/components/shadcn/input';
+import { Label } from '@/components/shadcn/label';
+import { Textarea } from '@/components/shadcn/textarea';
 import { Save, Loader2 } from 'lucide-react';
 import { useMaterial, useUpdateMaterial } from '@/hooks/materials';
 import { useCurrentUserEmployee } from '@/hooks/employee';
