+++ app/users/dashboard/resources/materials/page.tsx
@@ -2,9 +2,14 @@
 
 import { useMemo, useState } from 'react';
 import Link from 'next/link';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
+import {
+  Card,
+  CardContent,
+  CardHeader,
+  CardTitle,
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
+import { Badge } from '@/components/shadcn/badge';
 import { Pagination, SearchAndFilter } from '@/components/common';
 import {
   Select,
@@ -12,7 +17,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
+} from '@/components/shadcn/select';
 import { Package, Plus, Tag, Ruler, WarehouseIcon } from 'lucide-react';
 import { useMaterials } from '@/hooks/materials';
 import { MATERIAL_UNITS } from '@/features/materials/components/material-unit-selector';
